refactor chart type config constants
