+++ app/constants/chartTypeConfigs/discreteBarChart.js
@@ -1,0 +1,8 @@
+export const config = {
+  type: 'discreteBarChart',
+  label: 'Bar Chart',
+  dataFormat: 'nvd3SingleSeries',
+  componentName: 'DiscreteBarChart',
+};
+
+export const defaultOpts = {};
