+++ app/constants/chartTypeConfigs/lineChart.js
@@ -1,0 +1,8 @@
+export const config = {
+  type: 'lineChart',
+  label: 'Line Chart',
+  dataFormat: 'nvd3MultiSeries',
+  componentName: 'LineChart',
+};
+
+export const defaultOpts = {};
