+++ app/constants/chartTypeConfigs/pieChart.js
@@ -1,0 +1,12 @@
+export const config = {
+  type: 'pieChart',
+  label: 'Pie Chart',
+  dataFormat: 'nvd3SingleSeries',
+  componentName: 'PieChart',
+};
+
+export const defaultOpts = {
+  donut: true,
+  showLegend: false,
+  showLabels: false,
+};
