+++ app/constants/chartTypes.js
@@ -1,70 +1,8 @@
 /**
  * Chart types registry
  */
-import update from 'react-addons-update';
+import * as discreteBarChart from './chartTypeConfigs/discreteBarChart';
+import * as lineChart from './chartTypeConfigs/lineChart';
+import * as pieChart from './chartTypeConfigs/pieChart';
 
-export const selectableChartTypes = [
-  {
-    type: 'pieChart',
-    label: 'Pie Chart',
-    dataFormat: 'nvd3SingleSeries',
-    componentName: 'PieChart',
-    defaultOpts: {
-      donut: true,
-      showLegend: false,
-      showLabels: false,
-    },
-  },
-  {
-    type: 'discreteBarChart',
-    label: 'Bar Chart',
-    dataFormat: 'nvd3SingleSeries',
-    componentName: 'DiscreteBarChart',
-  },
-  {
-    type: 'lineChart',
-    label: 'Line Chart',
-    dataFormat: 'nvd3MultiSeries',
-    componentName: 'LineChart',
-  },
-];
-
-const nvd3Defaults = {
-  nvd3SingleSeries: {
-    x: (point) => point.label,
-    y: (point) => point.value,
-    height: 400,
-  },
-  nvd3MultiSeries: {
-    x: (point) => point.x,
-    y: (point) => point.y,
-    height: 400,
-  },
-};
-
-export function getChartTypeObject(type) {
-  for (const typeObj of selectableChartTypes) {
-    if (type === typeObj.type) {
-      return typeObj;
-    }
-  }
-  return null;
-}
-
-export function getChartTypeDefaultOpts(type) {
-  const typeObj = getChartTypeObject(type);
-  if (0 === typeObj.dataFormat.indexOf('nvd3')) {
-    // merge chart type into data format defaults
-    let returnOpts = update(nvd3Defaults[typeObj.dataFormat],
-      { $merge: { type: typeObj.type } });
-    // add any specific defaults if present
-    if (typeObj.defaultOpts) {
-      returnOpts = update(returnOpts, { $merge: typeObj.defaultOpts });
-    }
-    return returnOpts;
-  }
-  /**
-   * non-NVD3 data formats would do something else here
-   */
-  return {};
-}
+export const selectableChartTypes = [discreteBarChart, pieChart, lineChart];
